Add new stucture
